Start refactoring project
