+++ GUI/scanned_card.txt
@@ -1,0 +1,1 @@
+19764094
