+++ GUI/templates/Themes/Standard.css
@@ -17,7 +17,7 @@ header {
 }
 
 main {
-  background-image: url("/GUI/Themes/Theme-Images/BarStools.jpg");
+  background-image: url("/templates/Themes/Theme-Images/BarStools.jpg");
   background-size: 1500px 750px;
   margin-top: 1%;
   margin-left: 1%;
