+++ GUI/templates/layout.html
@@ -21,7 +21,7 @@
         <menu>
           <ul>
             <a href="{{ url_for('home') }}"><li id="Setup">Setup</li></a>
-            <a href="{{ url_for('menu') }}"><li id="Menu">Menu</li></a>
+            <a href="{{ url_for('menu.menu') }}"><li id="Menu">Menu</li></a>
             {% if  session['Auth'] %}
               {% if session['Auth'] == True %}
                   <a href="{{ url_for('setting') }}"><li id="Settings">Settings</li></a>
