+++ GUI/templates/upload-menu.html
