--- Init/Users.csv
@@ -1,3 +1,0 @@
-ID, Credit, Name, Access
-19760047, 77, Tom, Admin
-
